jarvis/requirements.txt
# JARVIS Requirements
# Install with: pip install -r requirements.txt

# Voice Recognition
SpeechRecognition>=3.10.0
PyAudio>=0.2.13

# Text-to-Speech
pyttsx3>=2.90

# System monitoring
psutil>=5.9.0

# Document processing (optional but recommended)
python-docx>=0.8.11
PyPDF2>=3.0.0
Pillow>=9.0.0
striprtf>=0.0.26

# Screen capture and OCR
pytesseract>=0.3.10
mss>=9.0.0

# Note: For OCR to work, you also need to install Tesseract-OCR:
# Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
# macOS: brew install tesseract
# Linux: sudo apt-get install tesseract-ocr

# AI/LLM integrations (optional - add your preferred one)
# openai>=1.0.0
# anthropic>=0.7.0

# Other utilities
requests>=2.28.0
